Move some types out of the ast module
