+++ base/src/error.rs
@@ -1,10 +1,11 @@
 //! Module containing a few common error wrappers which allows more information to be saved for
 //! later display to the user
+
 use std::any::Any;
 use std::error::Error as StdError;
 use std::fmt;
 
-use ast::Spanned;
+use pos::Spanned;
 
 /// An error type which can represent multiple errors.
 #[derive(Debug, PartialEq)]
