+++ src/import.rs
@@ -1,4 +1,5 @@
 //! Implementation of the `import` macro.
+
 use std::any::Any;
 use std::sync::{Arc, RwLock, Mutex};
 use std::fs::File;
@@ -8,6 +9,7 @@ use std::path::{Path, PathBuf};
 
 use base::ast;
 use base::metadata::Metadata;
+use base::pos;
 use base::symbol::Symbol;
 use vm::macros::{Macro, Error as MacroError};
 use vm::thread::{Thread, ThreadInternal};
@@ -163,7 +165,7 @@ impl<I> Macro for Import<I>
                     try!(self.importer.import(vm, &modulename, file_contents));
                 }
                 // FIXME Does not handle shadowing
-                Ok(ast::located(arguments[0].location,
+                Ok(pos::located(arguments[0].location,
                                 ast::Expr::Identifier(TcIdent::new(name))))
             }
             _ => return Err(Error::String("Expected a string literal to import".into()).into()),
